bt-740: * refactor
