+++ mobile/src/bundles/chat/components/components.ts
@@ -3,6 +3,5 @@ export { ChatHeader } from './chat-header/chat-header';
 export { ChatInfoButton } from './chat-info-button/chat-info-button';
 export { ChatItem } from './chat-item/chat-item';
 export { ChatListItem } from './chat-list-item/chat-list-item';
-export { CompanyInfo } from './company-info/company-info';
 export { MessageEntryField } from './message-entry-field/message-entry-field';
 export { Search } from './search/search';
