+++ mobile/src/bundles/chat/screens/chat-user-details/styles.ts
