+++ mobile/src/bundles/common/types/navigation/employer-bottom-tab-navigation-parameter-list.type.ts
@@ -2,7 +2,7 @@ import { type EmployerBottomTabScreenName } from '~/bundles/common/enums/enums';
 
 type EmployerBottomTabNavigationParameterList = {
     [EmployerBottomTabScreenName.CANDIDATES]: undefined;
-    [EmployerBottomTabScreenName.MAIL]: undefined;
+    [EmployerBottomTabScreenName.CHAT_ROOT]: undefined;
     [EmployerBottomTabScreenName.EMPLOYER_PROFILE]: undefined;
 };
